Refactor test specs.
