+++ test/EventHandlers/Pool/Sync.test.ts
@@ -1,8 +1,8 @@
 import { expect } from "chai";
-import { Pool, MockDb } from "../../generated/src/TestHelpers.gen";
-import { LiquidityPoolAggregator } from "../../generated/src/Types.gen";
+import { Pool, MockDb } from "../../../generated/src/TestHelpers.gen";
+import { LiquidityPoolAggregator } from "../../../generated/src/Types.gen";
 import { setupCommon } from "./common";
-import { TEN_TO_THE_18_BI, TEN_TO_THE_6_BI, toChecksumAddress, TokenIdByChain } from "../../src/Constants";
+import { TEN_TO_THE_18_BI, TEN_TO_THE_6_BI, toChecksumAddress, TokenIdByChain } from "../../../src/Constants";
 
 describe("Pool Sync Event", () => {
   let mockToken0Data: any;
